style: css 수정

+++ front/src/hooks/auth/useDeleteId.tsx
@@ -5,11 +5,18 @@ import Page_Url from "../../router/Url";
 function useDeleteId() {
   const navigate = useNavigate();
   const deleteId = () => {
+    const isConfirmed = window.confirm("정말 탈퇴하시겠습니까?");
+    if (!isConfirmed) {
+      return;
+    }
+
     deleteMember().then(() => {
       localStorage.removeItem("oauthProvider");
-      localStorage.removeItem("accessToken");
-      localStorage.removeItem("refreshToken");
-      localStorage.removeItem("uid");
+    localStorage.removeItem("accessToken");
+    localStorage.removeItem("refreshToken");
+    localStorage.removeItem("uid");
+    localStorage.removeItem("visitedPerfumes");
+    localStorage.removeItem("nickname");
       navigate(Page_Url.Landing);
     });
   };
